+++ src/main/java/fr/eni/PizzaOnline/bll/CustomerManagerImpl.java
@@ -15,30 +15,38 @@ public class CustomerManagerImpl implements CustomerManager{
 	CustomerDAO customerDAO;
 	
 	@Override
+	// TEST : OK
 	public void addCustomer(Customer customer) {
 		customerDAO.save(customer);
 		
 	}
 
 	@Override
+	// TEST : NOK
 	public void modCustomer(Customer customer) {
 		customerDAO.save(customer);
 	}
 
 	@Override
-	public void delCustomer(Long id) {
+	// TEST : OK pour customer sans commande
+	// test à réaliser avec commande
+	public void delCustomer(Integer id) {
 		Customer customer = getCustomerById(id);
 		customerDAO.delete(customer);
 		
 	}
 
 	@Override
+	// TEST : NOK
+	// failed to lazily initialize a collection of role: fr.eni.PizzaOnline.bo.Customer.pizzaOrder: could not initialize proxy - no Session
 	public List<Customer> getAllCustomers() {
 		return (List<Customer>) customerDAO.findAll();
 	}
 
 	@Override
-	public Customer getCustomerById(Long id) {
+	// TEST : NOK
+	// failed to lazily initialize a collection of role: fr.eni.PizzaOnline.bo.Customer.pizzaOrder: could not initialize proxy - no Session
+	public Customer getCustomerById(Integer id) {
 		return customerDAO.findById(id).orElse(null);
 	}
