+++ src/main/java/fr/eni/PizzaOnline/dal/BaseDAO.java
@@ -3,6 +3,6 @@ package fr.eni.PizzaOnline.dal;
 import org.springframework.data.repository.CrudRepository;
 import fr.eni.PizzaOnline.bo.Base;
 
-public interface BaseDAO extends CrudRepository<Base, Long>{
+public interface BaseDAO extends CrudRepository<Base, Integer>{
 
 }
