+++ src/main/java/fr/eni/PizzaOnline/dal/CustomerDAO.java
@@ -3,6 +3,6 @@ package fr.eni.PizzaOnline.dal;
 import org.springframework.data.repository.CrudRepository;
 import fr.eni.PizzaOnline.bo.Customer;
 
-public interface CustomerDAO extends CrudRepository<Customer, Long>{
+public interface CustomerDAO extends CrudRepository<Customer, Integer>{
 
 }
