+++ src/main/java/fr/eni/PizzaOnline/dal/FoodDAO.java
@@ -3,6 +3,6 @@ package fr.eni.PizzaOnline.dal;
 import org.springframework.data.repository.CrudRepository;
 import fr.eni.PizzaOnline.bo.Food;
 
-public interface FoodDAO extends CrudRepository<Food, Long>{
+public interface FoodDAO extends CrudRepository<Food, Integer>{
 
 }
