+++ src/main/java/fr/eni/PizzaOnline/dal/PizzaDAO.java
@@ -3,6 +3,6 @@ package fr.eni.PizzaOnline.dal;
 import org.springframework.data.repository.CrudRepository;
 import fr.eni.PizzaOnline.bo.Pizza;
 
-public interface PizzaDAO extends CrudRepository<Pizza, Long>{
+public interface PizzaDAO extends CrudRepository<Pizza, Integer>{
 
 }
