+++ src/main/java/fr/eni/PizzaOnline/dal/PizzaOrderDAO.java
@@ -3,7 +3,7 @@ package fr.eni.PizzaOnline.dal;
 import org.springframework.data.repository.CrudRepository;
 import fr.eni.PizzaOnline.bo.PizzaOrder;
 
-public interface PizzaOrderDAO extends CrudRepository<PizzaOrder, Long>{
+public interface PizzaOrderDAO extends CrudRepository<PizzaOrder, Integer>{
 
 }
